Make platform constraints more flexible
I've got toolchains for yocto based linux distributions which are very
much linux and unix, but don't match the constraints here.  I want to be
able to build network tables for them.  Really, the goal of these is to
detect linux, not the particular blessed variants of linux, so say that
in the constraints.

Signed-off-by: Austin Schuh <austin.linux@gmail.com>